Update to SDK 0.2.0 and openstorage-sdk-clients
This changes makes it much easier to update since it depends on
openstorage-sdk-clients. Running dep ensure is much cleaner.

To do this, the internal unit test had to be removed, and in
its place we use a docker container to run the mock-sdk-server.
This can now be run using `make test`.

A new Makefile was added to the root of the directory to make it
easier to run tests and build sdk-test.

Signed-off-by: Luis Pabón <luis@portworx.com>